// src/app/model/discover/questionLadder.model.ts
export class QuestionLadderModel {
  who: {
    is: string;
    did: string;
    can: string;
    will: string;
    would: string;
    might: string;
  };
  what: {
    is: string;
    did: string;
    can: string;
    will: string;
    would: string;
    might: string;
  };
  when: {
    is: string;
    did: string;
    can: string;
    will: string;
    would: string;
    might: string;
  };
  where: {
    is: string;
    did: string;
    can: string;
    will: string;
    would: string;
    might: string;
  };
  why: {
    is: string;
    did: string;
    can: string;
    will: string;
    would: string;
    might: string;
  };
  how: {
    is: string;
    did: string;
    can: string;
    will: string;
    would: string;
    might: string;
  };
}